chore: update code structure
