+++ packages/utils/src/index.ts
@@ -1,6 +1,5 @@
 export * from './store/node'
 export * from './store/idb'
-export * from './redux'
 export * from './tools/common'
 export * from './tools/tool'
 export * from './polyfill/recover-native'
@@ -8,5 +7,4 @@ export * from './tools/dom'
 export * from './tools/scroll'
 export * from './tools/output'
 export * from './tools/transform'
-export * from './performance/fmp'
 export * from './performance/radix64'
